requirements.txt
# Core dependencies
openai>=1.12.0
anthropic>=0.18.0
python-dotenv>=1.0.0

# OAuth and HTTP
requests>=2.31.0
flask>=3.0.0
requests-oauthlib>=1.3.1

# Data validation and parsing
pydantic>=2.5.0
jsonschema>=4.20.0

# Utilities
python-dateutil>=2.8.2
colorama>=0.4.6

# Testing (optional)
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development (optional)
black>=23.12.0
flake8>=6.1.0
mypy>=1.7.0